CLI skeleton framework to support native and def cluster commands
