+++ container_service_extension/client/vsphere_kubernetes.py
@@ -1,0 +1,13 @@
+# container-service-extension
+# Copyright (c) 2020 VMware, Inc. All Rights Reserved.
+# SPDX-License-Identifier: BSD-2-Clause
+
+from pyvcloud.vcd.exceptions import OperationNotSupportedException
+
+
+class VsphereKubernetes:
+    """Embedded Kubernetes into vSphere."""
+
+    def __getattr__(self, name):
+        msg = "Operation not supported; Under implementation"
+        raise OperationNotSupportedException(msg)
